+++ CHANGES.rst
@@ -170,6 +170,8 @@ astropy.stats
 
 astropy.table
 ^^^^^^^^^^^^^
+- ``Table.from_pandas`` now supports a ``units`` dictionary as argument to pass units
+  for columns in the ``DataFrame``. [#9472]
 
 astropy.tests
 ^^^^^^^^^^^^^
@@ -850,7 +852,7 @@ astropy.uncertainty
 astropy.units
 ^^^^^^^^^^^^^
 
-- Support for unicode parsing. Currently supported superscripts are Ohm,
+- Support for unicode parsing. Currently supported are superscripts, Ohm,
   Ångström, and the micro-sign. [#9348]
 
 - Accept non-unit type annotations in @quantity_input. [#8984]